feat: style light theme
