+++ frontend/src/components/CustomButton/customButton.config.ts
@@ -1,6 +1,8 @@
 export const styles = {
   primary:
     "text-white bg-dark-400 px-8 py-4 font-bold rounded-lg hover:bg-dark-50 dark:bg-gray-100 dark:text-dark-400 dark:hover:bg-white",
-  transparent: "text-dark-400 dark:text-gray-200",
-  flattened: "bg-blue-100 px-4 rounded-lg py-2 text-white dark:bg-blue-500",
+  transparent:
+    "text-dark-400 dark:text-gray-200 hover:text-dark-50 dark:hover:text-white",
+  flattened:
+    "bg-blue-100 px-4 rounded-lg py-2 text-white dark:bg-blue-500 hover:bg-blue-500 dark:hover:bg-blue-100",
 };
